bakalari_strava_app/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
requests
flask_migrate
strava-cz